Refactor corpus construction.
Uses a different sparse matrix structure to improve speed and remove
C++11 dependency. Some other changes to make the package pypi-ready.

+++ tests/utils.py
@@ -55,4 +55,5 @@ def build_coocurrence_matrix(sentences):
 
     return sp.coo_matrix((data, (rows, cols)),
                          shape=(len(dictionary),
-                                len(dictionary))).tocsr().tocoo()
+                                len(dictionary)),
+                         dtype=np.float32).tocsr().tocoo()
